Add fixed with integer types
Add fixed with integer types like uint8 int32 etc, and had to create a mypy for to support automatic conversion from int literals to these fixed width ints, so things like i: uint8 = 10 work directly. Otherwise you would need i: uint8 = uint8(10)

+++ src/splice/runtime/types.h
@@ -1,7 +1,9 @@
 #pragma once
 
+#include <cstdint>
+
 // The C++ types representing Python's `int` and `float` - change here to
 // retarget the whole runtime. Global namespace, not py::, so unnamespaced
 // headers and generated code can use them unqualified.
-using _int = long long;
+using _int = int64_t;
 using _float = double;
